requirements.txt
vertexai
google-cloud-aiplatform
google-cloud-texttospeech
google-cloud-speech